get location from cookie
