+++ static/src/javascripts.flow.archive/projects/common/modules/commercial/support-utilities.spec.js
@@ -13,7 +13,7 @@
 // @flow
 
 import { addCountryGroupToSupportLink } from 'common/modules/commercial/support-utilities';
-import { setGeolocation } from 'lib/geolocation';
+import { setGeolocation } from 'static/src/javascripts/lib/geolocation';
 
 describe('addCountryGroupToSupportLink', () => {
     test('adds country group to subscribe link', () => {
