+++ static/src/javascripts/projects/common/modules/commercial/reader-revenue-dev-utils.js
@@ -1,7 +1,7 @@
 import { storage } from '@guardian/libs';
 import { addCookie, removeCookie } from '../../../../lib/cookies';
 import {
-    getSync as geolocationGetSync,
+    get as getCountryCode,
     setGeolocation,
 } from '../../../../lib/geolocation';
 import {
@@ -108,7 +108,7 @@ const showPreviousVariant = (asExistingSupporter = false) => {
 
 const changeGeolocation = (asExistingSupporter = false) => {
     const geo = window.prompt(
-        `Enter two-letter geolocation code (e.g. GB, US, AU). Current is ${geolocationGetSync()}.`
+        `Enter two-letter geolocation code (e.g. GB, US, AU). Current is ${getCountryCode()}.`
     );
     if (geo === 'UK') {
         alert(`'UK' is not a valid geolocation - please use 'GB' instead!`);
