+++ static/src/javascripts/projects/common/modules/commercial/support-utilities.spec.js
@@ -1,29 +1,39 @@
 import { addCountryGroupToSupportLink } from './support-utilities';
-import { setGeolocation } from '../../../../lib/geolocation';
+import { getCookie } from '../../../../lib/cookies';
+
+jest.mock('../../../../lib/cookies', () => ({
+    getCookie: jest.fn(() => null),
+}));
+
+jest.mock('lib/raven');
 
 describe('addCountryGroupToSupportLink', () => {
-    test('adds country group to subscribe link', () => {
-        setGeolocation('GB');
+    beforeEach(() => {
+        getCookie.mockImplementation(() => null);
+    });
+
+    test('adds country group to subscribe link', async () => {
+        getCookie.mockImplementation(() => 'GB');
         expect(
-            addCountryGroupToSupportLink(
+            await addCountryGroupToSupportLink(
                 'https://support.theguardian.com/subscribe'
             )
         ).toEqual('https://support.theguardian.com/uk/subscribe');
     });
 
-    test('adds country group to contribute link', () => {
-        setGeolocation('FR');
+    test('adds country group to contribute link', async () => {
+        getCookie.mockImplementation(() => 'FR');
         expect(
-            addCountryGroupToSupportLink(
+            await addCountryGroupToSupportLink(
                 'https://support.theguardian.com/contribute'
             )
         ).toEqual('https://support.theguardian.com/eu/contribute');
     });
 
-    test('does not add country group to contribute link with country group already in it', () => {
-        setGeolocation('GB');
+    test('does not add country group to contribute link with country group already in it', async () => {
+        getCookie.mockImplementation(() => 'GB');
         expect(
-            addCountryGroupToSupportLink(
+            await addCountryGroupToSupportLink(
                 'https://support.theguardian.com/int/contribute'
             )
         ).toEqual('https://support.theguardian.com/int/contribute');
