+++ static/src/javascripts/projects/common/modules/identity/sign-in-gate/helper.js
@@ -2,7 +2,6 @@ import bean from 'bean';
 import userPrefs from 'common/modules/user-prefs';
 import { storage } from '@guardian/libs';
 import config from 'lib/config';
-import { getSync as geolocationGetSync } from 'lib/geolocation';
 import {
     getSynchronousTestsToRun,
     isInABTestSynchronous,
@@ -164,13 +163,6 @@ export const isNPageOrHigherPageView = (n = 2) => {
 
     // check if count is greater or equal to 1 less than n since dailyArticleCount is incremented after this component is loaded
     return count >= n - 1;
-};
-
-// use gu.location to determine is the browser is in the specified country
-// Note, use country codes specified in /static/src/javascripts/lib/geolocation.js
-export const isCountry = (countryCode) => {
-    const geolocation = geolocationGetSync();
-    return geolocation === countryCode;
 };
 
 // determine if the useragent is running iOS 9 (known to be buggy for sign in flow)
